refactor: isolate WebSocket client
Now there's going to be several ping/pong transports, pull all related
WebSocket code into an isolated file.
